Changing all metrics calculation from float64 to float32

+++ codes/ranking_metrics_torch_karlhigley/common.py
@@ -16,8 +16,8 @@ def _check_inputs(ks, scores, labels):
 
     return (
         ks.to(dtype=torch.int32, device=scores.device),
-        scores.to(dtype=torch.float64, device=scores.device),
-        labels.to(dtype=torch.float64, device=scores.device),
+        scores.to(dtype=torch.float32, device=scores.device),
+        labels.to(dtype=torch.float32, device=scores.device),
     )
 
 
@@ -30,5 +30,5 @@ def _extract_topk(ks, scores, labels):
 
 def _create_output_placeholder(scores, ks):
     return torch.zeros(scores.shape[0], len(ks)).to(
-        device=scores.device, dtype=torch.float64
+        device=scores.device, dtype=torch.float32
     )
